// packages/lyrics-kit/src/service/types/netease/searchResult.d.ts
interface NetEaseResponseSong {
    name: string;
    id: number;
    /** duration (in millisecond) */
    duration: number;
    artists: {
        name: string;
        id: number;
    }[];
    album: {
        name: string;
        id: number;
        picUrl: string;
    };
}

interface NetEaseResponseSearchResult {
    result: {
        songCount: number;
        songs: NetEaseResponseSong[];
    };
    code: number;
}